1대다 전투 구현, 전투 행동 선택지 구현, 레벨업 처리
적 조우시 1대다 전투가 되도록 구현함
 - 적을 모두 처치시에만 경험치, 골드 획득
 - 공격 행동 선택지 공격할 적 선택하도록 함

플레이어 레벨업 처리
 - 플레이어 경험치 획득 시 레벨업 처리

그 외
 - 랜덤 엔진은 namespace를 만들어서 외부로 뺌
 - 마을에서 info 출력, 여관 회복 구현

+++ BaseGame/GameManager.cpp
@@ -118,7 +118,7 @@ void GameManager::Run()
 			
 			// 최상단 스택부터 검사해서 불투명한 씬을 찾고 찾았다면 break
 			int idx = 0;
-			for (int i = static_cast<int>(scene_stack.size()) - 1; i >= 0; --i) {
+			for (int i = scene_stack.size() - 1; i >= 0; --i) {
 				if (scene_stack[i]->IsOpaque()) {
 					idx = i;
 					break;
@@ -126,7 +126,7 @@ void GameManager::Run()
 			}
 
 			// 불투명한 씬부터 렌더링, 이 아래는 어차피 안보이므로 그릴 필요 X
-			for (int i = idx; i < static_cast<int>(scene_stack.size()); ++i) {
+			for (int i = idx; i < scene_stack.size(); ++i) {
 				scene_stack[i]->Render();
 				scene_stack[i]->RenderUI();
 			}
